feature(slim-api): renamed to `subquery`
