+++ packages/groqd/src/commands/subquery/conditional.test.ts
