+++ packages/groqd/src/commands/subquery/select.test.ts
