+++ packages/groqd/src/commands/subquery/select.ts
